+++ src/ansibleguy-webui/aw/static/js/jobs/repository.js
@@ -2,6 +2,9 @@ function updateApiTableDataRepository(row, entry) {
     row.innerHTML = document.getElementById('aw-api-data-tmpl-row').innerHTML;
     row.cells[0].innerText = entry.name;
     row.cells[1].innerText = entry.rtype_name;
+    let actionsTemplate = document.getElementById("aw-api-data-tmpl-actions").innerHTML;
+    row.cells[4].innerHTML = actionsTemplate.replaceAll('${ID}', entry.id);
+
     if (entry.rtype_name == "Git") {
         row.cells[2].innerHTML = entry.git_origin + ':' + entry.git_branch;
     } else {
@@ -9,6 +12,8 @@ function updateApiTableDataRepository(row, entry) {
     }
     if (entry.rtype_name == "Static") {
         row.cells[3].innerText = '-';
+        let actionButtonUpdate = row.cells[4].getElementsByClassName("aw-repo-update")[0];
+        actionButtonUpdate.setAttribute("hidden", "hidden");
     } else if (is_set(entry.status_name) && !entry.git_isolate) {
         if (is_set(entry.time_update)) {
             row.cells[3].innerHTML += '<b>Updated</b>: ' + entry.time_update + '<br>';
@@ -31,8 +36,6 @@ function updateApiTableDataRepository(row, entry) {
         row.cells[3].innerHTML += '<br>' + statusTemplate;
     }
 
-    let actionsTemplate = document.getElementById("aw-api-data-tmpl-actions").innerHTML;
-    row.cells[4].innerHTML = actionsTemplate.replaceAll('${ID}', entry.id);
 }
 
 $( document ).ready(function() {