+++ test/integration/webui/main.py
@@ -52,7 +52,7 @@ def test_get_locations(locations: list):
 
 def test_main_pages():
     test_get_locations([
-        'ui/jobs/manage', 'ui/jobs/log', 'ui/jobs/credentials',
+        'ui/jobs/manage', 'ui/jobs/log', 'ui/jobs/credentials', 'ui/jobs/repository',
         'ui/settings/api_keys', 'ui/settings/permissions',
         'ui/system/admin/', 'ui/system/api_docs', 'ui/system/environment', 'ui/system/config',
         'a/password_change/',
@@ -63,7 +63,7 @@ def test_actions_views():
     test_get_locations([
         'ui/jobs/manage/job',
         'ui/jobs/credentials/0?global=false', 'ui/jobs/credentials/0?global=true',
-        'ui/settings/permissions/0',
+        'ui/settings/permissions/0', 'ui/jobs/repository/git/0', 'ui/jobs/repository/static/0'
     ])