+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   * Select using Up/Down/Enter Keys
 * Ability to clone existing jobs
 * Ability to sort jobs on manage UI
+* Split-up Repository Forms
 
 ----